requirements.txt
python-telegram-bot==20.7

requests==2.32.5

flask==3.1.2